fix : elastic 위치 이동
