+++ src/main/java/com/nhnacademy/bookstore/global/elastic/controller/ElasticController.java
@@ -1,15 +1,12 @@
-package com.nhnacademy.bookstore.elastic.test;
+package com.nhnacademy.bookstore.global.elastic.controller;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.elasticsearch.core.SearchHits;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,33 +16,30 @@ import com.nhnacademy.bookstore.book.bookCartegory.service.BookCategoryService;
 import com.nhnacademy.bookstore.book.bookTag.dto.request.ReadBookIdRequest;
 import com.nhnacademy.bookstore.book.bookTag.dto.response.ReadTagByBookResponse;
 import com.nhnacademy.bookstore.book.bookTag.service.BookTagService;
-import com.nhnacademy.bookstore.elastic.book.repository.ElasticSearchBookRepository;
-import com.nhnacademy.bookstore.elastic.book.repository.ElasticSearchCustomBookRepository;
-import com.nhnacademy.bookstore.elastic.document.book.BookDocument;
+import com.nhnacademy.bookstore.global.elastic.book.repository.ElasticSearchBookRepository;
+import com.nhnacademy.bookstore.global.elastic.document.book.BookDocument;
+import com.nhnacademy.bookstore.util.ApiResponse;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @RestController
-public class ElasticTestController {
+@RequestMapping("/bookstore/books/search")
+@RequiredArgsConstructor
+public class ElasticController {
+	private final ElasticSearchBookRepository elasticSearchBookRepository;
+	private final BookService bookService;
+	private final BookCategoryService bookCategoryService;
+	private final BookTagService bookTagService;
 
-	@Autowired
-	ElasticSearchBookRepository elasticSearchBookRepository;
-
-	@Autowired
-	ElasticSearchCustomBookRepository elasticSearchCustomBookRepository;
-
-	@Autowired
-	BookService bookService;
-	@Autowired
-	BookCategoryService bookCategoryService;
-
-	@Autowired
-	BookTagService bookTagService;
-
-	@GetMapping("/test/push")
+	/**
+	 * 책 밀어 넣어주기
+	 * 나중에 배치서버로 이동시키거나 삭제할 예정
+	 * @return 성공 메시지
+	 */
+	@GetMapping("/push")
 	String push() {
-
 		for (long i = 1; i <= 713; i++) {
 			try {
 				ReadBookResponse book = bookService.readBookById(i);
@@ -82,19 +76,21 @@ public class ElasticTestController {
 		return "good";
 	}
 
-	@GetMapping("/test/search/{indexId}")
-	public BookDocument searchIndex(@PathVariable Long indexId) {
-		return elasticSearchBookRepository.findById(indexId).orElse(null);
+	/**
+	 * 검색 결과 보내주기
+	 * @param page 페이지
+	 * @param size    페이지 사이즈
+	 * @param keyword    검색 키워드
+	 * @return 검색으로 찾은 책들
+	 */
+	@GetMapping
+	public ApiResponse<Page<BookDocument>> searchKeyWord(@RequestParam(defaultValue = "0") int page,
+		@RequestParam(defaultValue = "12") int size,
+		@RequestParam("keyword") String keyword) {
+
+		Page<BookDocument> pageBookDocument = elasticSearchBookRepository.findByCustomQuery(keyword,
+			PageRequest.of(page, size));
+		return ApiResponse.success(pageBookDocument);
 	}
 
-	@GetMapping("/test/search/title/{title}")
-	public Page<BookDocument> searchTitle(@PathVariable String title) {
-		return elasticSearchBookRepository.findByCustomQuery(title, PageRequest.of(0, 10));
-	}
-
-	@GetMapping("/test/search/keyword/{keyword}")
-	public SearchHits<BookDocument> searchKeyword(@PathVariable String keyword, @RequestParam int page) {
-		Pageable pageable = PageRequest.of(page - 1, 10);
-		return elasticSearchCustomBookRepository.searchProductsByProductName(keyword, pageable);
-	}
 }
